requirements.txt
# NutriAI - requirements.txt
# Install: pip install -r requirements.txt

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9        # file uploads
python-dotenv==1.0.1

# ── Database ───────────────────────────────────────────────────────────────────
psycopg[binary]==3.2.3         # async PostgreSQL driver (psycopg3)

# ── LangChain + LangGraph ──────────────────────────────────────────────────────
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-chroma==0.1.4
langgraph==0.2.45
langgraph-checkpoint-postgres==2.0.5   # AsyncPostgresSaver

# ── LLM + Embeddings ───────────────────────────────────────────────────────────
huggingface-hub==0.25.2
sentence-transformers==3.2.1   # all-MiniLM-L6-v2 embeddings

# ── Vector store ───────────────────────────────────────────────────────────────
chromadb==0.5.18

# ── Computer vision ────────────────────────────────────────────────────────────
torch==2.4.1                   # EfficientNet-B2 inference
torchvision==0.19.1
timm==1.0.11                   # model architecture
Pillow==10.4.0                 # image preprocessing

# ── OCR ────────────────────────────────────────────────────────────────────────
pytesseract==0.3.13            # Tesseract fallback (requires system binary)
# Google Vision API uses httpx — no extra package needed

# ── HTTP client ────────────────────────────────────────────────────────────────
httpx==0.27.2                  # HuggingFace API calls, Vision API

# ── Fuzzy matching ─────────────────────────────────────────────────────────────
rapidfuzz==3.10.0              # dish name fuzzy matching

# ── Observability (optional) ───────────────────────────────────────────────────
langsmith==0.1.143             # LangSmith tracing (set LANGCHAIN_TRACING_V2=true)

# ── Utilities ──────────────────────────────────────────────────────────────────
pydantic==2.9.2
